+++ src/main/java/racingcar/domain/Cars.java
@@ -1,23 +1,35 @@
-package racingcar;
+package racingcar.domain;
 
+import racingcar.controller.RacingCar;
+
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class Cars {
 
+    public static final String COMMA = ",";
     private List<Car> carList;
+
+    public Cars(String carNameInput) {
+        this.carList = new ArrayList<>();
+        String[] carNameArr = carNameInput.split(COMMA);
+        for (String carName : carNameArr) {
+            carList.add(new Car(carName));
+        }
+    }
 
     public Cars(List<Car> carList) {
         this.carList = carList;
     }
 
-    protected void move() {
+    public void move() {
         for (Car car : carList) {
             car.move(RacingCar.getRandom());
         }
     }
 
-    List<Car> findWinner() {
+    public List<Car> findWinner() {
         int maxDistance = getMaxDistance(this.carList);
 
         return carList.stream().filter(car -> car.distance().matchDistance(maxDistance)).collect(Collectors.toList());
@@ -32,7 +44,7 @@ public class Cars {
         return maxDistance;
     }
 
-    List<Car> Cars() {
+    public List<Car> getCars() {
         return this.carList;
     }
