+++ src/main/java/racingcar/view/ResultView.java
@@ -1,18 +1,21 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 
 import java.util.List;
 
 public class ResultView {
     public static final String COLON = " : ";
     public static final String HYPHEN = "-";
-    public static final String COMMA = ", ";
+    public static final String COMMA_SPACE = ", ";
 
     public static void output(String msg){
         System.out.println(msg);
     }
 
     public static void resultOutput(Cars cars) {
-        List<Car> carList = cars.Cars();
+        List<Car> carList = cars.getCars();
         for (Car car : carList) {
             System.out.println(car.name() + COLON + HYPHEN.repeat(car.distance().distance()));
         }
@@ -22,7 +25,7 @@ public class ResultView {
         String result = "";
         for (Car car : carList) {
             result += car.name();
-            result += COMMA;
+            result += COMMA_SPACE;
         }
         result = removeLastComma (result);
         result += "가 최종 우승했습니다";
